modified clock, new field, more comments, validation for numbers, installed faker

+++ address.py
@@ -1,6 +1,9 @@
 import tkinter as tk
 from tkinter import messagebox
 import time
+from faker import Faker
+
+fake = Faker()
 
 def hora():
     """Funcion para mostrar la hora actual"""
@@ -11,17 +14,30 @@ def hora():
 def agregar_contacto():
     """Funcion para agregar un contacto"""
     nombre = entrada_nombre.get()
+    apellido = entrada_apellido.get()
     telefono = entrada_telefono.get()
     direccion = entrada_direccion.get()
     
-    if nombre and telefono and direccion:
+    if nombre and telefono and direccion and apellido:
         lista_contactos.insert(tk.END, nombre)
-        contactos[nombre] = {"teléfono": telefono, "dirección": direccion}
+        contactos[nombre] = {"apellido": apellido,"teléfono": telefono, "dirección": direccion}
         entrada_nombre.delete(0, tk.END)
+        entrada_apellido.delete(0, tk.END)
         entrada_telefono.delete(0, tk.END)
         entrada_direccion.delete(0, tk.END)
     else:
         messagebox.showwarning("Campos vacíos", "Por favor, completa todos los campos antes de agregar un contacto.")
+
+def agregar_contactos_falsos():
+    """Usando Faker creamos unos usuarios"""
+    cantidad = 10
+    for _ in range(cantidad):
+        nombre = fake.name()
+        apellido = fake.last_name()
+        telefono = fake.phone_number().replace("x", " ")
+        dir = fake.address().replace("\n", ", ")
+        lista_contactos.insert(tk.END, nombre)
+        contactos[nombre] = {"apellido": apellido,"teléfono": telefono, "dirección": dir}
 
 def eliminar_contacto():
     """Función para eliminar un contacto"""
@@ -43,12 +59,13 @@ def mostrar_info_contacto():
     nombre = lista_contactos.get(seleccion[0])
     contacto = contactos.get(nombre)
 
-    # Crear una ventana modal
+    # Modal
     modal_ventana = tk.Toplevel(ventana)
     modal_ventana.title("Información Detallada del Contacto")
-    modal_ventana.geometry('300x200')
+    modal_ventana.geometry('400x200')
 
-    tk.Label(modal_ventana, text=f"Nombre: {nombre}").pack(side='top', anchor='nw',padx=(0, 10), pady=5)
+    tk.Label(modal_ventana, text=f"Nombre: {nombre} {contacto['apellido']}").pack(side='top', anchor='nw',padx=(0, 10), pady=5)
+    # tk.Label(modal_ventana, text=f"Apellido: {contacto['apellido']}").pack(side='top', anchor='nw',padx=(0, 10), pady=5)
     tk.Label(modal_ventana, text=f"Teléfono: {contacto['teléfono']}").pack(side='top', anchor='nw',padx=(0, 10), pady=5)
     tk.Label(modal_ventana, text=f"Dirección: {contacto['dirección']}").pack(side='top', anchor='nw',padx=(0, 10), pady=5)
 
@@ -64,6 +81,11 @@ def limpiar_lista():
     lista_contactos.delete(0, tk.END)
     contactos.clear()
 
+# Validaciones
+def validar_telefono(texto):
+    """validar que se permitan solo numeros o espacios en blanco"""
+    return True if texto == "" or texto.replace("-","").isdigit() else False
+
 # Crear la ventana principal
 ventana = tk.Tk()
 ventana.title('Agenda de Contactos')
@@ -78,11 +100,11 @@ menu_principal.add_command(label='Limpiar lista', command=limpiar_lista)
 menu_principal.add_command(label='Salir', command=salir)
 
 # reloj en la parte superior derecha
-reloj = tk.Label(ventana, font=('Arial', 20), fg='black')
+reloj = tk.Label(ventana, font=('Arial', 10), fg='black')
 reloj.place(relx=1.0, rely=0.0, anchor='ne')
 hora()
 
-# Marco para el boton de nombre
+# Marco para el boton de nombre y apellido
 marco_nombre = tk.Frame(ventana)
 marco_nombre.pack(anchor='nw', padx=10, pady=(10, 0))
 
@@ -90,14 +112,21 @@ marco_nombre.pack(anchor='nw', padx=10, pady=(10, 0))
 # Crear las entradas para los detalles del contacto
 tk.Label(marco_nombre, text="Nombre:").pack(side='left', padx=(0, 10))
 entrada_nombre = tk.Entry(marco_nombre)
-entrada_nombre.pack(side='left')
+entrada_nombre.pack(side='left', padx=(0, 10))
+
+tk.Label(marco_nombre, text="Apellido:").pack(side='left', padx=(2, 10))
+entrada_apellido = tk.Entry(marco_nombre)
+entrada_apellido.pack(side='left', padx=(6, 10))
 
 # Marco para los botones de telefono y direccion
 marco_contacto = tk.Frame(ventana)
 marco_contacto.pack(anchor='nw', padx=10, pady=(10, 0))
 
+# Registro de validacion de numero
+v_telefono = ventana.register(validar_telefono)
+
 tk.Label(marco_contacto, text="Teléfono:").pack(side='left', padx=(0, 10))
-entrada_telefono = tk.Entry(marco_contacto)
+entrada_telefono = tk.Entry(marco_contacto, validate="key", validatecommand=(v_telefono, '%P'))
 entrada_telefono.pack(side='left', padx=(0,10))
 
 tk.Label(marco_contacto, text="Dirección:").pack(side='left', padx=(0, 10))
@@ -128,6 +157,10 @@ boton_eliminar.pack(side='left', anchor='nw', padx=10, pady=(0,10))
 # Boton para mostrar la información detallada del contacto
 boton_mostrar_info = tk.Button(marco_botones_abajo, text='Mostrar Info', command=mostrar_info_contacto)
 boton_mostrar_info.pack(side='left', anchor='nw', padx=(0, 10))
+
+# Boton para agregar dummy info
+boton_fake = tk.Button(marco_botones_abajo, text='F data', command=agregar_contactos_falsos)
+boton_fake.pack(side='left', anchor='nw', padx=(0, 10))
 
 # Iniciar el bucle de la ventana
 ventana.mainloop()
